+++ common/GameConfig.cpp
@@ -54,6 +54,7 @@ bool GameConfig::load() try
     result &= dash_faction_key.read_value("Keep Launcher Open", &keep_launcher_open);
     result &= dash_faction_key.read_value("Skip Cutscene Control", &skip_cutscene_ctrl);
     result &= dash_faction_key.read_value("Screen Flash", &screen_flash);
+    result &= dash_faction_key.read_value("Language", &language);
 
 #ifdef NDEBUG
     max_fps = std::clamp(max_fps, MIN_FPS_LIMIT, MAX_FPS_LIMIT);
@@ -111,6 +112,7 @@ void GameConfig::save() try
     dash_faction_key.write_value("Keep Launcher Open", keep_launcher_open);
     dash_faction_key.write_value("Skip Cutscene Control", skip_cutscene_ctrl);
     dash_faction_key.write_value("Screen Flash", screen_flash);
+    dash_faction_key.write_value("Language", language);
 }
 catch (...) {
     std::throw_with_nested(std::runtime_error("failed to save config"));