+++ launcher/resource.h
@@ -44,6 +44,7 @@
 #define IDC_KEEP_LAUNCHER_OPEN_CHECK    1033
 #define IDC_HIGH_MON_RES_CHECK          1034
 #define IDC_LINEAR_PITCH_CHECK          1035
+#define IDC_LANG_COMBO                  1036
 
 // Next default values for new objects
 //
@@ -51,7 +52,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        104
 #define _APS_NEXT_COMMAND_VALUE         32771
-#define _APS_NEXT_CONTROL_VALUE         1036
+#define _APS_NEXT_CONTROL_VALUE         1037
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif